requirements.txt
# Core Requirements
llama-cpp-python==0.2.32
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
pydantic==2.5.3

# Testing & Analytics
colorama==0.4.6
requests==2.31.0
matplotlib==3.7.1
seaborn==0.12.2

# Optional Enhancements
python-dotenv==1.0.0
aiofiles==23.2.1